changed loat to double in time scale

+++ src/ECS/entitymanager.hpp
@@ -189,7 +189,7 @@ public:
         m_timeScale = timeScale;
     }
 
-    float getTimeScale() {
+    double getTimeScale() {
         return m_timeScale;
     }
